# JsonGenerator/endpoint-server/openapi/swaggerInit.yaml
...
openapi: 3.1.0
info:
  title: GeoRDFBench Framework JSON Specification API with Swagger
  version: 2.0.0
  description: A cRud API application made with Express and documented with Swagger
  license:
    name: Apache-2.0
    url: https://spdx.org/licenses/Apache-2.0.html
  contact:
    name: Theofilos Ioannidis
    url: http://users.uoa.gr/~tioannid/
    email: tioannid@di.uoa.gr
servers:
  - url: http://localhost:{port}/
    description: Development server
    variables:
      port:
        enum:
          - "5000"
          - "3000"
        default: "5000"
  - url: "{scheme}://geordfbench.di.uoa.gr/jsonapi/"
    description: Production server
    variables:
      scheme:
        enum:
          - http
          - https
        default: http
paths:
...
  "/hosts/{spec}":
    get:
      summary: Get the host by specification name
      tags:
        - Hosts
      parameters:
        - in: path
          name: spec
          schema:
            type: string
          required: true
          description: The host specification name
      responses:
        "200":
          description: The specification for the host specification name provided
          contens:
            application/json:
              schema:
                "$ref": "#/components/schemas/Host"
        "404":
          description: The host specification name was not found
  "/hosts/{newspec}":
    post:
      summary: Create the host by specification name
      requestBody:
        description: Specify a **JSON** string
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Host"
      tags:
        - Hosts
      parameters:
        - in: path
          name: newspec
          schema:
            type: string
          required: true
          description: The host specification name
      responses:
        "201":
          description: The host specification was created
          contens:
            application/json:
              schema:
                "$ref": "#/components/schemas/Host"
        "400":
          description: A field is either missing or the ram field is not integer!
        "404":
          description: The host specification name already exists!
  "/hosts/{existingspec}":
    delete:
      summary: Delete an existing host by specification name
      tags:
        - Hosts
      parameters:
        - in: path
          name: existingspec
          schema:
            type: string
          required: true
          description: The host specification name
      responses:
        "204":
          description: The host specification was deleted
        "404":
          description: The host specification name was not found
...
  "/querysets/{spec}":
    get:
      summary: Get the queryset by specification name
      tags:
        - Querysets
      parameters:
        - in: path
          name: spec
          schema:
            type: string
          required: true
          description: The queryset specification name
      responses:
        "200":
          description: The specification for the queryset specification name provided
          contens:
            application/json:
              schema:
                "$ref": "#/components/schemas/Queryset"
        "404":
          description: The queryset specification name was not found
  "/querysets/{newspec}":
    post:
      summary: Create the queryset by specification name
      requestBody:
        description: Specify a **JSON** string
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Queryset"
      tags:
        - Querysets
      parameters:
        - in: path
          name: newspec
          schema:
            type: string
          required: true
          description: The queryset specification name
      responses:
        "201":
          description: The queryset specification was created
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Queryset"
        "400":
          description: SimpleQueryset fields are missing!
        "403":
          description: StaticTempParamQS or DynamicTempParamQueryset specific fields are missing!
        "404":
          description: The queryset specification name already exists!
  "/querysets/{existingspec}":
    delete:
      summary: Delete an queryset by specification name
      tags:
        - Querysets
      parameters:
        - in: path
          name: existingspec
          schema:
            type: string
          required: true
          description: The queryset specification name
      responses:
        "204":
          description: The queryset specification was deleted
        "404":
          description: The queryset specification name was not found
...
components:
  schemas:
...
    Os:
      type: object
      properties:
        classname:
          type: string
          description: The java full class name
        name:
          type: string
          description: The name of the operating system
        shell_cmd:
          type: string
          description: The command line interpreter
        sync_cmd:
          type: string
          description: The command used to synchronize cached data to permanent storage
        clearcache_cmd:
          type: string
          description: The command to clear system caches
    Host:
      type: object
      required:
        - classname
        - name
        - ipAddr
        - ram
        - os
        - sourceFileDir
        - reposBaseDir
        - reportsBaseDir
      properties:
        classname:
          type: string
          description: The java full class name
        name:
          type: string
          description: The name of the host
        ipAddr:
          type: string
          format: ipv4
          description: The host IP4 address
        ram:
          type: number
          format: int32
          description: Host's total RAM in GBs
        os:
          "$ref": "#/components/schemas/Os"
        sourceFileDir:
          type: string
          format: path
          description: Host's base location for dataset files
        reposBaseDir:
          type: string
          format: path
          description: Host's base location for RDF store repositories
        reportsBaseDir:
          type: string
          format: path
          description: Host's base location for experiment reports
      example:
        classname: gr.uoa.di.rdf.geordfbench.runtime.hosts.impl.SimpleHost
        name: NUC8i7BEH
        ipAddr: 192.168.1.44
        ram: 32
        os:
          classname: gr.uoa.di.rdf.geordfbench.runtime.os.impl.UbuntuJammyOS
          name: Ubuntu-jammy
          shell_cmd: "/bin/sh"
          sync_cmd: sync
          clearcache_cmd: sudo /sbin/sysctl vm.drop_caches=3
        sourceFileDir: "/data/Geographica2_Datasets"
        reposBaseDir: "/data"
        reportsBaseDir: "/data/Results_Store"
...
    SimpleQueryset:
      type: object
      required:
        - classname
        - name
        - relativeBaseDir
        - hasPredicateQueriesAlso
        - mapQueries
        - mapUsefulNamespacePrefixes
      properties:
        classname:
          type: string
          description: The java full class name
        name:
          type: string
          description: The name of the queryset
        relativeBaseDir:
          type: string
          format: path
          description: The queryset's path relative to the host's sourceFileDir
        hasPredicateQueriesAlso:
          type: boolean
          description: There exist queries with spatial predicate queries
        mapQueries:
          type: array
          description: The queries that make up the queryset
        mapUsefulNamespacePrefixes:
          type: array
          description: The queries that make up the queryset
      example:
        classname: gr.uoa.di.rdf.geordfbench.runtime.querysets.complex.impl.SimpleQS
        name: scalabilityFunc
        relativeBaseDir: ""
        hasPredicateQueriesAlso: false
        mapQueries:
          "0":
            label: SC1_Geometries_Intersects_GivenPolygon
            text: "SELECT ?s1 ?o1 WHERE { \n ?s1 geo:asWKT ?o1 . \n  FILTER(geof:sfIntersects(?o1, \"POLYGON((23.708496093749996 37.95719224376526,22.906494140625 40.659805938378526, 23.708496093749996 37.95719224376526))\"^^<http://www.opengis.net/ont/geosparql#wktLiteral>)). \n} \n"
            usePredicate: false
            expectedResults: -1
        mapUsefulNamespacePrefixes: {}
    StaticTempParamQueryset:
      allOf:
        - $ref: "#/components/schemas/SimpleQueryset"
          type: object
          required:
            - mapTemplateParams
            - mapGraphPrefixes
          properties:
            mapTemplateParams:
              type: array
              description: The template parameters used in the queryset
            mapGraphPrefixes:
              type: array
              description: The named graph prefixes closely related to the target dataset
          example:
            classname: gr.uoa.di.rdf.geordfbench.runtime.querysets.complex.impl.StaticTempParamQS
            name: scalabilityFunc
            relativeBaseDir: ""
            hasPredicateQueriesAlso: false
            mapQueries:
              "0":
                label: SC1_Geometries_Intersects_GivenPolygon
                text:
                  "SELECT ?s1 ?o1 WHERE { \n ?s1 geo:asWKT ?o1 . \n  FILTER(geof:FUNCTION(?o1,
                  GIVEN_SPATIAL_LITERAL)). \n} \n"
                usePredicate: false
                expectedResults: -1
              "1":
                label: SC2_Intensive_Geometries_Intersect_Geometries
                text:
                  "SELECT ?s1 ?s2 \nWHERE { \n ?s1 geo:hasGeometry [ geo:asWKT ?o1
                  ] ;\n    lgo:has_code \"1001\"^^xsd:integer . \n ?s2 geo:hasGeometry
                  [ geo:asWKT ?o2 ] ;\n    lgo:has_code ?code2 .  \n FILTER(?code2>5000
                  && ?code2<6000 && ?code2 != 5260) .\n FILTER(geof:FUNCTION(?o1, ?o2)).
                  \n} \n"
                usePredicate: false
                expectedResults: -1
              "2":
                label: SC3_Relaxed_Geometries_Intersect_Geometries
                text:
                  "SELECT ?s1 ?s2\nWHERE {\n ?s1 geo:hasGeometry [ geo:asWKT ?o1 ]
                  ;\n    lgo:has_code \"1001\"^^xsd:integer .\n ?s2 geo:hasGeometry [
                  geo:asWKT ?o2 ] ;\n    lgo:has_code ?code2 .\n FILTER(?code2 IN (5622,
                  5601, 5641, 5621, 5661)) .\n FILTER(geof:FUNCTION(?o1, ?o2)).\n} \n"
                usePredicate: false
                expectedResults: -1
            mapUsefulNamespacePrefixes: {}
            mapTemplateParams:
              GIVEN_SPATIAL_LITERAL:
                '"POLYGON((23.708496093749996 37.95719224376526,22.906494140625
                40.659805938378526,11.524658203125002 48.16425348854739,-0.1181030273437499
                51.49506473014367,-3.2189941406250004 55.92766341247031,-5.940856933593749
                54.59116279530599,-3.1668090820312504 51.47967237816337,23.708496093749996
                37.95719224376526))"^^<http://www.opengis.net/ont/geosparql#wktLiteral>'
              FUNCTION: sfIntersects
            mapGraphPrefixes: {}
    DynamicTempParamQueryset:
      allOf:
        - $ref: "#/components/schemas/SimpleQueryset"
          type: object
          required:
            - mapQueryTemplates
            - mapLiteralValues
            - templateDynamicParamNameList
            - templateParamValueMatrix
            - isTranslated
          properties:
            mapQueryTemplates:
              type: object
              description: The template name or template query
              additionalProperties:
                type: string
            mapLiteralValues:
              type: object
              description: The map of named literal values (static template parameters common for the entire queryset)
              additionalProperties:
                type: string
            templateDynamicParamNameList:
              type: array
              description: The list of template parameter names
            templateParamValueMatrix:
              type: object
              description: The map of query number and dynamic template parameter pairs to values
              additionalProperties:
                type: string
            isTranslated:
              type: boolean
              description: Translation of template parameters should be done only once
          example:
            classname: gr.uoa.di.rdf.geordfbench.runtime.querysets.complex.impl.DynamicTempParamQS
            name: rwmicro
            relativeBaseDir: ""
            hasPredicateQueriesAlso: false
            mapQueries:
              "0":
                label: Buffer_GeoNames
                text: NONTOP_BUFFER
                usePredicate: false
                expectedResults: 21990
              "1":
                label: Buffer_LGD
                text: NONTOP_BUFFER
                usePredicate: false
                expectedResults: 12097
              "2":
                label: Within_GeoNames_Point_Buffer
                text: SELECT_BUFFER
                usePredicate: false
                expectedResults: 152
              "3":
                label: Touches_GAG_GAG
                text: JOIN_2
                usePredicate: false
                expectedResults: 1022
            mapUsefulNamespacePrefixes:
              uom: "<http://www.opengis.net/def/uom/OGC/1.0/>"
              strdf: "<http://strdf.di.uoa.gr/ontology#>"
            mapQueryTemplates:
              SELECT_BUFFER: "SELECT ?s1 \nWHERE {\n\tGRAPH <GRAPH1> {\n\t\t?s1 <ASWKT1> ?o1\n\t}\nFILTER(geof:<FUNCTION1>(?o1, geof:<FUNCTION2>(\"POINT(23.71622 37.97945)\"^^<http://www.opengis.net/ont/geosparql#wktLiteral>, 3000, uom:metre))).\n}"
              NONTOP_BUFFER: "SELECT (geof:<FUNCTION1>(?o1, 4, uom:metre) AS ?ret) \nWHERE {\n\tGRAPH <GRAPH1> {\n\t\t?s1 <ASWKT1> ?o1\n\t}\n}"
              JOIN_2: "SELECT ?s1 ?s2 \nWHERE { \n\tGRAPH <GRAPH1> {\n\t\t?s1 <ASWKT1> ?o1\n\t}\n\tGRAPH <GRAPH2> {\n\t\t?s2 <ASWKT2> ?o2\n\t}\nFILTER(?s1 != ?s2).\nFILTER(geof:<FUNCTION1>(?o1, ?o2)).\n}"
            mapLiteralValues:
              LL_clc: "<http://geographica.di.uoa.gr/dataset/clc>"
              LL_clc_asWKT: "<http://geo.linkedopendata.gr/corine/ontology#asWKT>"
              LL_gadm_asWKT: "<http://geo.linkedopendata.gr/gag/ontology/asWKT>"
              LL_geonames_asWKT: "<http://www.geonames.org/ontology#asWKT>"
              LL_fn_area: area
              LL_sfWithin: sfWithin
              LL_lgd: "<http://geographica.di.uoa.gr/dataset/lgd>"
              LL_gadm: "<http://geographica.di.uoa.gr/dataset/gag>"
              LL_sfTouches: sfTouches
              LL_geonames: "<http://geographica.di.uoa.gr/dataset/geonames>"
              LL_lgd_asWKT: "<http://linkedgeodata.org/ontology/asWKT>"
              LL_fn_buffer: buffer
            templateDynamicParamNameList:
              - <FUNCTION1>
              - <FUNCTION2>
              - <ASWKT1>
              - <ASWKT2>
              - <GRAPH1>
              - <GRAPH2>
            templateParamValueMatrix:
              mapTemplateParamValues:
                "0##<GRAPH1>": LL_geonames
                "1##<GRAPH1>": LL_lgd
                "2##<FUNCTION2>": LL_fn_buffer
                "2##<FUNCTION1>": LL_sfWithin
                "1##<ASWKT1>": LL_lgd_asWKT
                "3##<FUNCTION1>": LL_sfTouches
                "2##<GRAPH1>": LL_geonames
                "2##<ASWKT1>": LL_geonames_asWKT
                "3##<GRAPH2>": LL_gadm
                "3##<GRAPH1>": LL_gadm
                "3##<ASWKT2>": LL_gadm_asWKT
                "3##<ASWKT1>": LL_gadm_asWKT
                "1##<FUNCTION1>": LL_fn_buffer
                "0##<ASWKT1>": LL_geonames_asWKT
                "0##<FUNCTION1>": LL_fn_buffer
            isTranslated: false
...
    Queryset:
      oneOf:
        - $ref: "#/components/schemas/SimpleQueryset"
        - $ref: "#/components/schemas/StaticTempParamQueryset"
        - $ref: "#/components/schemas/DynamicTempParamQueryset"
        - $ref: "#/components/schemas/MacroGeocodingQueryset"
        - $ref: "#/components/schemas/MacroMapSearchQueryset"
...